添加组件工厂类，EntitySystem类，现在组件的创建操作以及组件运行能够均匀地工作在各个线程中，实现负载均衡

+++ src/libs/libserver/util_time.h
@@ -7,12 +7,11 @@
 
 namespace timeutil
 {
-
 	typedef uint64 Time;  // milliseconds from 1971
 
 	inline Time AddSeconds(Time timeValue, int second)
 	{
-		return timeValue + second * 1000;
+		return timeValue + (uint64)second * 1000;
 	}
 
 	inline Time AddMilliseconds(Time timeValue, int milliseconds)
